scripts/requirements.txt
google-cloud-storage>=2.14.0
google-auth>=2.28.0
